Sketch out `pqxx::prepped` marker type.
This is a new way to call prepared statements.  Instead of encoding
the fact that it's a prepared statement in the function name, you'll
wrap the name in a lightweight object which marks it as a prepared
statement name.

I'm also hoping to extract the row count checks (`exec0()`, `exec1()`
etc.) into a separate helper, so that most of the calls to the various
"exec" calls will just be to functions called `exec()`.  Whether they're
calls to prepared statements, to parameterised statements, or plain SQL
will become more of an implementation detail that the library derives
from the arguments.